Minor OpenAFS client changes.
Fix use return value of VOP_ISLOCKED in a couple of spots, although I
don't think this was an issue.

Also, skip vinvalbuf when we're being called as a pager. I'm not sure
what the consequences of this are, but vinvalbuf will surely deadlock
the thread if it's called in a pager: it waits for paging to finish, but
it IS the pager, so it's stuck.